+++ xivo_sysconf/modules/asterisk.py
@@ -90,11 +90,6 @@ def _is_valid_path_component(path_component):
                 os.sep not in path_component)
 
 
-def core_show_version(args, options):
-    command = 'core show version'
-    return _exec_asterisk_command(command)
-
-
 def core_show_channels(args, options):
     command = 'core show channels'
     return _exec_asterisk_command(command)
@@ -117,7 +112,5 @@ http_json_server.register(asterisk.delete_voicemail, CMD_R,
                           name='delete_voicemail')
 http_json_server.register(asterisk.move_voicemail, CMD_R,
                           name='move_voicemail')
-http_json_server.register(core_show_version, CMD_R,
-                          name='core_show_version')
 http_json_server.register(core_show_channels, CMD_R,
                           name='core_show_channels')