+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 -----
 
 * The Asterisk command `sip show peer <peer>` has been removed since it does not work with PJSIP and is not used
+* The Asterisk command `core show version` has been removed since it is not used anymore
 
 
 18.14